Remove dark hex game state.
Address comments in #15.
