+++ lib/setup.py
@@ -8,10 +8,8 @@ setup(
     scripts=['bin/games_puzzles_algorithms_gtp.py'],
     install_requires=[
         'future == 0.15.2',
-        'setuptools == 20.2.2',
-        "cffi >= 1.0.0"
+        'setuptools == 20.2.2'
     ],
     tests_require=['pytest'],
-    setup_requires=["cffi >= 1.0.0", 'pytest-runner'],
-    cffi_modules=["games_puzzles_algorithms/games/dex/hyper_cube_indexer_build.py:ffi"]
+    setup_requires=['pytest-runner']
 )
